+++ src/pynwb/ndx_microscopy/testing/__init__.py
@@ -2,6 +2,7 @@ from ._mock import (
     mock_MicroscopyRig,
     mock_MicroscopeModel,
     mock_Microscope,
+    mock_MicroscopyChannel,
     mock_Segmentation,
     mock_Segmentation2D,
     mock_Segmentation3D,
@@ -22,6 +23,7 @@ from ._mock import (
 __all__ = [
     "mock_Microscope",
     "mock_MicroscopeModel",
+    "mock_MicroscopyChannel",
     "mock_MicroscopyRig",
     "mock_PlanarImagingSpace",
     "mock_VolumetricImagingSpace",